requirements.txt
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
av==17.0.0
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.1
cloup==3.0.8
decorator==5.2.1
distro==1.9.0
glcontext==3.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ImageIO==2.37.3
imageio-ffmpeg==0.6.0
isosurfaces==0.1.2
jiter==0.13.0
manim==0.20.1
ManimPango==0.6.1
mapbox_earcut==2.0.0
markdown-it-py==4.0.0
mdurl==0.1.2
moderngl==5.12.0
moderngl-window==3.1.1
moviepy==2.2.1
networkx==3.6.1
numpy==2.4.4
openai==2.30.0
outcome==1.3.0.post0
packaging==26.0
pillow==11.3.0
proglog==0.1.12
pycairo==1.29.0
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
pyglet==2.1.13
pyglm==2.8.3
Pygments==2.20.0
PySocks==1.7.1
python-dotenv==1.2.2
requests==2.33.1
rich==14.3.3
scipy==1.17.1
screeninfo==0.8.1
selenium==4.41.0
skia-pathops==0.9.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
srt==3.5.3
svgelements==1.9.6
tqdm==4.67.3
trio==0.33.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
watchdog==6.0.0
webdriver-manager==4.0.2
websocket-client==1.9.0
wsproto==1.3.2
